Switch to upstream uAPI definition
Extract local API into a BSD-licensed copy of the upstream
misc/fastrpc.h header.
