+++ inc/remote.h
@@ -6,6 +6,7 @@
 
 #include <stdint.h>
 #include <sys/types.h>
+#include <misc/fastrpc.h>
 
 #ifdef __cplusplus
 extern "C" {
@@ -646,62 +647,6 @@ enum remote_mem_map_flags {
 
 /** Update REMOTE_MAP_MAX_FLAG when adding new value to this enum **/
  };
-
-/**
- * @enum fastrpc_map_flags for fastrpc_mmap and fastrpc_munmap
- * @brief Types of maps with cache maintenance
- */
-enum fastrpc_map_flags {
-    /**
-     * Map memory pages with RW- permission and CACHE WRITEBACK.
-     * Driver will clean cache when buffer passed in a FastRPC call.
-     * Same remote virtual address will be assigned for subsequent
-     * FastRPC calls.
-     */
-    FASTRPC_MAP_STATIC,
-
-    /** Reserved for compatibility with deprecated flag */
-    FASTRPC_MAP_RESERVED,
-
-    /**
-     * Map memory pages with RW- permission and CACHE WRITEBACK.
-     * Mapping tagged with a file descriptor. User is responsible for
-     * maintenance of CPU and DSP caches for the buffer. Get virtual address
-     * of buffer on DSP using HAP_mmap_get() and HAP_mmap_put() functions.
-     */
-    FASTRPC_MAP_FD,
-
-    /**
-     * Mapping delayed until user calls HAP_mmap() and HAP_munmap()
-     * functions on DSP. User is responsible for maintenance of CPU and DSP
-     * caches for the buffer. Delayed mapping is useful for users to map
-     * buffer on DSP with other than default permissions and cache modes
-     * using HAP_mmap() and HAP_munmap() functions.
-     */
-    FASTRPC_MAP_FD_DELAYED,
-
-    /** Reserved for compatibility **/
-    FASTRPC_MAP_RESERVED_4,
-    FASTRPC_MAP_RESERVED_5,
-    FASTRPC_MAP_RESERVED_6,
-    FASTRPC_MAP_RESERVED_7,
-    FASTRPC_MAP_RESERVED_8,
-    FASTRPC_MAP_RESERVED_9,
-    FASTRPC_MAP_RESERVED_10,
-    FASTRPC_MAP_RESERVED_11,
-    FASTRPC_MAP_RESERVED_12,
-    FASTRPC_MAP_RESERVED_13,
-    FASTRPC_MAP_RESERVED_14,
-    FASTRPC_MAP_RESERVED_15,
-
-    /**
-     * This flag is used to skip CPU mapping,
-     * otherwise behaves similar to FASTRPC_MAP_FD_DELAYED flag.
-     */
-    FASTRPC_MAP_FD_NOMAP,
-
-    /** Update FASTRPC_MAP_MAX when adding new value to this enum **/
-};
 
 /**
  * Attributes for remote_register_buf_attr
