+++ src/fastrpc_apps_user.c
@@ -150,14 +150,7 @@ inline static void deinit_fastrpc_dsp_lib_refcnt(void) {
   }
 }
 
-enum fastrpc_proc_attr {
-  FASTRPC_MODE_DEBUG = 0x1,
-  FASTRPC_MODE_PTRACE = 0x2,
-  FASTRPC_MODE_CRC = 0x4,
-  FASTRPC_MODE_UNSIGNED_MODULE = 0x8,
-  FASTRPC_MODE_ADAPTIVE_QOS = 0x10,
-  FASTRPC_MODE_SYSTEM_PROCESS = 0x20,
-  FASTRPC_MODE_PRIVILEGED = 0x40, // this attribute will be populated in kernel
+enum fastrpc_proc_attr_local {
   // Attribute to enable pd dump feature for both signed/unsigned pd
   FASTRPC_MODE_ENABLE_PDDUMP = 0x80,
   // System attribute to enable pd dump debug data collection on rooted devices
